Changed logo URL to an absolute path
This is an include and the tags will be rendered in different contexts. 

The logo was broken on all pages besides the root due to the use of a relative path in an include. Changed to relative path so that the logo is displayed regardless of where this include is used.

```diff
-<img src="images/cognifide-logo.png"
+<img src="/images/cognifide-logo.png"
```